Add random letters
Now we have letters that change every 50 milliseconds. This required some new randoms to be added along with a new class and jquery (my love but also the bane of my aesthetic code wishes). Randomly generated fonts and letters added. Need to see if there's some garbage collection that can be done.